Gallery Lock - Flask/requirements.txt
alembic==1.8.1
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
filestack-python==3.5.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
pycparser==2.21
requests==2.28.1
SQLAlchemy==1.4.40
trafaret==2.0.2
urllib3==1.26.11
Werkzeug==2.2.2
WTForms==3.0.1
